refactor(database): 调整数据库模块位置、调整初始化arg名称，并更新文档
- 将 db.py 数据库模块移入‘database/‘中，并修正所有调用代码
- 将初始化参数 `--init-db` 改为 `--init_db`
- 更新相关文档说明
