+++ README.md
@@ -196,21 +196,29 @@ python main.py --help
 ## 💾 数据保存
 
 支持多种数据存储方式：
-
-- **SQLite 数据库**：轻量级数据库，无需服务器，适合个人使用（推荐）
-  - 参数：`--save_data_option sqlite`
-  - 自动创建数据库文件
-- **MySQL 数据库**：支持关系型数据库 MySQL 中保存（需要提前创建数据库）
-  - 执行 `python db.py` 初始化数据库表结构（只在首次执行）
 - **CSV 文件**：支持保存到 CSV 中（`data/` 目录下）
 - **JSON 文件**：支持保存到 JSON 中（`data/` 目录下）
+- **数据库存储**
+  - 使用参数 `--init_db` 进行数据库初始化（使用`--init_db`时不需要携带其他optional）
+  - **SQLite 数据库**：轻量级数据库，无需服务器，适合个人使用（推荐）
+    1. 初始化：`--init_db sqlite`
+    2. 数据存储：`--save_data_option sqlite`
+  - **MySQL 数据库**：支持关系型数据库 MySQL 中保存（需要提前创建数据库）
+    1. 初始化：`--init_db mysql`
+    2. 数据存储：`--save_data_option db`（db 参数为兼容历史更新保留）
+
 
 ### 使用示例：
 ```shell
-# 使用 SQLite（推荐个人用户使用）
+# 初始化 SQLite 数据库（使用'--init_db'时不需要携带其他optional）
+uv run main.py --init_db sqlite
+# 使用 SQLite 存储数据（推荐个人用户使用）
 uv run main.py --platform xhs --lt qrcode --type search --save_data_option sqlite
-
-# 使用 MySQL
+```
+```shell
+# 初始化 MySQL 数据库
+uv run main.py --init_db mysql
+# 使用 MySQL 存储数据（为适配历史更新，db参数进行沿用）
 uv run main.py --platform xhs --lt qrcode --type search --save_data_option db
 ```
